requirements.txt
neo-boa>=0.5.0
git+https://github.com/CityOfZion/neo-python.git@development#egg=neo-python